This is a HUGE MESS. Also, a bit of a step backwards.
However, I think we are on a better track. We now scroll through the
gutter instead of the left pane. We make the gutter as tall as all of the
tallest sides of the diff chunks combined. Additionally we maintain an
index of alignment points (left and right line numbers that should line up).
The index is keyed on gutter line number. So, as we scroll through the
gutter we can easily see what line number on the left should be matched
to which line number on the right for the current scroll position.

Still a ways off making this work. Need to figure out which pane to shift
up/down for a given alignment point. Need to make it smooth again.

Aaaaand, desperately need to clean up all of the sad sad dead code.