Added sub-diff highlighting! Contrast now highlights which parts of changed lines have actually been modified. This was notably more complicated to add than it would have otherwise been because of syntax highlighting.
The lines are already made up of many spans to highlight different logical
bits of code. Adding span wrapping around changed bits requires walking
both the diff chunks and the dom nodes at the same time and carefully
keeping track of where we are in each and how they correspond.

One issue is that the diff library I am using produces LOTS of little diff
chunks for sub diff (often individual characters) and this is not helpful.
I will need to see if there is a different way to use the library or a
different library altogether that I should be using.