Improve scrolling performance by taking over scrolling and using animation frames.
Previously we reacted to scrolling to adjust the left/right offset. This caused a jitter when scrolling because the offset adjustments would apply slightly after the content scrolled. Now, the offset adjustments occur at the same time as the scrolling. Further we make these changes via requestAnimationFrame() which helps avoid overwhelming rendering.