Track the active diff component instead of relying on :hover query.
Tracking the component instead of the element gives access to data on the component. This is needed to better scope the scrolling logic to individual file-diff components. Each file-diff component now binds to mouseenter/leave and emits activate/deactivate events. These events are observed by the parent components and used to set activeDiff on the app. Move the scroll listener into the app and use app.activeDiff to determine which file-diff component to manipulate when scrolling.

Reduce the opacity of sub-diff chunk highlights to improve readability on comments. Rename Vue.js instance to 'app'.