Fix a bug where active diff doesn't update unless mouse moves.
We were picking the active diff to scroll through based on mouseenter/leave events. This doesn't work well because those events don't fire unless the mouse moves. When scrolling the mouse might not move and therefore the active diff to scroll might not update. Now we loop through all diffs to find the ones that are visible based on scroll position and offset. Then we invoke our scrolling logic on all of them.